[FEAT] Toggle display of streets and weights
Added two buttons in the lower right corner to toggle the display of streets and weights on the map. It is now possible to hide and show both elements of independently. Flags ('showWeights' and 'showStreets') were implemented to control the conditional rendering of the streets and weights. The buttons update their state when pressed and reflect the changes on the screen.