[FEAT] Weight change due to traffic
Logic was added so that the weight of the streets can be changed according to the type of traffic there is. The traffic is defined by a product, if the route weighs 5, you can define normal traffic 1, moderate traffic 2, slow traffic 3. The route increases its weight with the product weight 5x3 = 15 (new weight).